+++ dpnp/__init__.py
@@ -70,11 +70,15 @@ from .dpnp_iface import __all__ as _iface__all__
 from .dpnp_iface_utils import *
 from .dpnp_iface_utils import __all__ as _ifaceutils__all__
 from ._version import get_versions
+from . import linalg as linalg
 from . import scipy as scipy
 
 __all__ = _iface__all__
 __all__ += _ifaceutils__all__
 __all__ += ["scipy"]
+
+# add submodules
+__all__ += ["linalg"]
 
 
 __version__ = get_versions()["version"]