+++ dpnp/dpnp_iface.py
@@ -52,7 +52,6 @@ import dpnp
 from dpnp.dpnp_algo import *
 from dpnp.dpnp_array import dpnp_array
 from dpnp.fft import *
-from dpnp.linalg import *
 from dpnp.memory import *
 from dpnp.random import *